Address Codex review: validate serialization format at FFI boundary
- FFI: `ferric_engine_serialize_as` and `ferric_engine_deserialize_as` now
  accept `u32` instead of `FerricSerializationFormat` enum, with explicit
  validation returning `FERRIC_ERROR_INVALID_ARGUMENT` for out-of-range
  discriminants. Eliminates UB from C callers passing invalid enum values.

- Go: `formatToFFI` now returns an error for unknown formats instead of
  silently defaulting to bincode, consistent with `toFFIConflictStrategy`
  and `toFFIStringEncoding`.

- Includes remaining unstaged work from the multi-format serde branch:
  CLI snapshot command, REPL snapshot restore, Python file-based
  serialization, Go file helpers and tests.

+++ bindings/go/internal/ffi/serialization.go
@@ -15,7 +15,7 @@ func EngineSerializeAs(h EngineHandle, format SerializationFormat) ([]byte, Erro
 	var data *C.uint8_t
 	var length C.uintptr_t
 
-	rc := ErrorCode(C.ferric_engine_serialize_as(h, format, nil, nil, &data, &length)) //nolint:gocritic // dupSubExpr false positive in cgo-generated code
+	rc := ErrorCode(C.ferric_engine_serialize_as(h, C.uint32_t(format), nil, nil, &data, &length)) //nolint:gocritic // dupSubExpr false positive in cgo-generated code
 	if rc != ErrOK {
 		return nil, rc
 	}
@@ -46,7 +46,7 @@ func EngineDeserializeAs(data []byte, format SerializationFormat) (EngineHandle,
 	rc := ErrorCode(C.ferric_engine_deserialize_as(
 		(*C.uint8_t)(unsafe.Pointer(&data[0])),
 		C.uintptr_t(len(data)),
-		format,
+		C.uint32_t(format),
 		&engine, //nolint:gocritic // dupSubExpr false positive in cgo-generated code
 	))
 	return engine, rc
